Refactor stacked bar widget
